# Add Windows icon support to pywebview