@r0x0r 2025-06-27
The icon for a Windows app is extracted from the exe and bundled by a freezing app. How does your solution interact with frozen app?