@r0x0r 2025-06-28
@gitcnd macOS use an icon from the executable, there is no separate icon implementation. Linux implementation uses an external icon, because as far as I know this is the only way to get an application icon. Windows can use both, but in my opinion bundling an icon in an executable is a correct way to do it. If both ways are supported, it would mean you would need to bundle an icon in an executable package separately, instead of doing it in a native way.